Fix: add shared Amplify test setup
